Fix double jump
