+++ Assets/Prefabs/Player.prefab
@@ -78,7 +78,6 @@ GameObject:
   - component: {fileID: 3941042754428619202}
   - component: {fileID: 2274185813217286231}
   - component: {fileID: 1319459820}
-  - component: {fileID: 1574135090300674427}
   - component: {fileID: 5723027543902039374}
   - component: {fileID: 1338750286670374244}
   m_Layer: 11
@@ -157,6 +156,7 @@ MonoBehaviour:
   AirMovementPenalty: 0.75
   damageFromFallMinDistance: 10
   fallDamageMult: 0.2
+  isApplyingFallDamage: 1
 --- !u!114 &251117981135048682
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -222,7 +222,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 4806390366692392982}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fff0190872d260c4e8aae4eebf4bb654, type: 3}
+  m_Script: {fileID: 11500000, guid: 20728bad22ce7d8498f13f3ed0ec86a3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   MAX_DJUMP: 1
@@ -242,7 +242,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Speed: 20
   MaxGrappleDistance: 100
---- !u!114 &1574135090300674427
 --- !u!114 &5723027543902039374
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -258,12 +257,6 @@ MonoBehaviour:
   ClimbSpeed: 10
   CanClimbDistance: 1.5
   ClimbDecayMultiplier: 1.2
-  m_Script: {fileID: 11500000, guid: 8aac66593db3ef14c97b5e85ee13aa21, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  moveDirection: {x: 0, y: 0, z: 0}
-  dashDistance: 10
-  dashStoppingSpeed: 0.1
 --- !u!114 &1338750286670374244
 MonoBehaviour:
   m_ObjectHideFlags: 0
