--- Assets/Scripts/Player/DoubleJump.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fff0190872d260c4e8aae4eebf4bb654
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
